Fix controller codestyle problems

+++ controller/lib/Timing/Stopwatch.h
@@ -1,17 +1,22 @@
+// Copyright 2022, Mikhael Khrustik <misha@myrt.co>
+//
+// All components of Myrt Desk Firmware are licensed under the BSD 3-Clause
+// License. See LICENSE.txt for details.
+
 #pragma once
 
 #include <Arduino.h>
 
 class Stopwatch {
-  public:
-    void start() {
-      _start = millis();
-    }
+ public:
+  void start() {
+    _start = millis();
+  }
 
-    size_t split() {
-      return millis() - _start;
-    }
+  size_t split() {
+    return millis() - _start;
+  }
 
-  private:
-    unsigned long _start;
+ private:
+  uint32_t _start;
 };
